# fix out-dated dependencies