perf improved a little:

old:

```
done 6578 ms
done 6589 ms
done 6644 ms
done 6657 ms
done 6684 ms
done 6688 ms
done 6695 ms
done 6777 ms
n                     : 4000
time in ms total      : 6777
time in ms per captcha: 1.69425
#captchs per second   : 590
```

after updated:

```
done 5677 ms
done 5736 ms
done 5737 ms
done 5745 ms
done 5761 ms
done 5765 ms
done 5798 ms
done 5810 ms
n                     : 4000
time in ms total      : 5810
time in ms per captcha: 1.4525
#captchs per second   : 688
```